Use contextFilter for graph modifications
